Defense Tests, Actions, Armor
Added a display of Defense Tests plus the optional Block/Dodge actions to the display mode and updated the action mode for the same.
Added in a display for major and minor action count to the initiative row for action/display modes.
Updated the armor values as well as the armor displays; augments that affect Defense Rating should be corrected now.